refactor cgroup build so non-linux systems see plugin
also updated the README for the fields->files change.
