+++ plugins/inputs/cgroup/cgroup_notlinux.go
@@ -1,3 +1,11 @@
 // +build !linux
 
 package cgroup
+
+import (
+	"github.com/influxdata/telegraf"
+)
+
+func (g *CGroup) Gather(acc telegraf.Accumulator) error {
+	return nil
+}
